pyproject.toml
[tool.poetry]
name = "repl-nix-profwebbuilder"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
psycopg2-binary = "^2.9.9"
wtforms = "^3.1.2"
requests = "^2.32.3"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
